feat: update sorting of course extension ideas based on reverse sort position for vote page
The sorting of course extension ideas in the `orderedCourseExtensionIdeas` computed property has been updated to use the `reverseSortPositionForVotePage` property instead of `createdAt`. This change ensures that the course extension ideas are sorted correctly based on their reverse sort position for the vote page. Additionally, the `reverseSortPositionForVotePage` property has been added to the `CourseExtensionIdea` model to calculate the reverse sort position based on the development status and creation date. This change improves the ordering and presentation of course extension ideas on the vote page.
